pyproject.toml
[tool.poetry]
name = "floppy-player"
version = "0.1.0"
description = ""
authors = ["Tau <hey@tau.garden>"]

[tool.poetry.dependencies]
python = "^3.11"
pyudev = "^0.24.1"
dbus-python = "^1.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
